[AIRFLOW-5254] Move GCP Tasks to core (#5860)
This commit moves GCP Tasks from contrib to core.
For more information check AIP-21.
